Refactor auxiliary classes
