+++ src/games/strategy/engine/framework/IServerRemote.java
@@ -1,0 +1,7 @@
+package games.strategy.engine.framework;
+
+import games.strategy.engine.message.IRemote;
+
+public interface IServerRemote extends IRemote {
+  byte[] getSavedGame();
+}
